+++ 01_src/com_r3dcomments/administrator/sql/updates/mysql/5.3.16.sql
@@ -1,2 +1,2 @@
 ALTER TABLE `#__r3dcomments`
-    ADD COLUMN IF NOT EXISTS `ip_hash` CHAR(64) NULL AFTER `ip`;
+    ADD COLUMN `ip_hash` CHAR(64) NULL AFTER `ip`;